Starting work on Fika Support
- Started working on Fika Support
- Starting introduction of a RaidManager class to handle active raids and active profiles
