+++ Server/src/constant.ts
@@ -1,0 +1,20 @@
+import config from '../config.json';
+
+export const WebSocketConfig = {
+    port: config.web_socket_port || 7828,
+    perMessageDeflate: {
+      zlibDeflateOptions: {
+        chunkSize: 1024,
+        memLevel: 7,
+        level: 3,
+      },
+      zlibInflateOptions: {
+        chunkSize: 10 * 1024,
+      },
+      clientNoContextTakeover: true,
+      serverNoContextTakeover: true,
+      serverMaxWindowBits: 10,
+      concurrencyLimit: 10,
+      threshold: 1024,
+    },
+}
